pyproject.toml
[project]
name = "downloads-organizer"
version = "0.1.0"
description = "Agent de tri automatique du dossier Downloads"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "httpx>=0.27",
    "ollama>=0.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-tmp-files>=0.0.2",
]

[project.scripts]
organize = "organizer.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
